requirements.txt
glfw>=2.6.4
PyOpenGL>=3.1.7
PyOpenGL_accelerate>=3.1.7
numpy>=1.24
